requirements.txt
aiogram==3.3.0
apscheduler==3.10.4
httpx==0.26.0
python-dotenv==1.0.0
pytz==2024.1 